apps/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "petehome-cli"
version = "2.0.0"
description = "Interactive TUI for petehome development management"
requires-python = ">=3.11"
dependencies = [
    "textual>=3.0.0",
    "textual-autocomplete>=4.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.28.0",
]

[project.scripts]
petehome = "petehome_cli.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
